requirements.txt
Django>=5.0
djangorestframework>=3.14
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
drf-spectacular>=0.27.0

# Docker
*.log
postgres_data/

djangorestframework-simplejwt>=5.3.0